requirements.txt
mysql-connector-python==8.0.32
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
Flask==3.0.3
Flask-Login==0.6.3